+++ arm_inlines.h
@@ -172,14 +172,14 @@ __INLINE DWORD gray2bin(const DWORD gray) {
 /***  VAR SIZE CONVERSIONS  ***/
 /******************************/
 
-/*!\brief converts 16bits to 8bits
+/*!\brief convert 16bits variable to 8bits
 ** \param[in] val - 16b value to convert
 ** \return Converted value
 **/
 __INLINE BYTE INLINE__ conv16to8Bits(const WORD val) {
 	return (BYTE) RSHIFT(val, 8); }
 
-/*!\brief converts 8bits to 16bits
+/*!\brief convert 8bits variable to 16bits
 ** \param[in] val - 8b value to convert
 ** \return Converted value
 **/
@@ -187,7 +187,37 @@ __INLINE WORD INLINE__ conv8to16Bits(const BYTE val) {
 	return (WORD) (LSHIFT(val, 8) + val); }
 
 
-/*!\brief converts 8bits to 8+nb bits (16bits max)
+/*!\brief convert 32bits variable to 16bits
+** \param[in] val - 32b value to convert
+** \return Converted value
+**/
+__INLINE WORD INLINE__ conv32to16Bits(const DWORD val) {
+	return (WORD) RSHIFT(val, 16); }
+
+/*!\brief convert 16bits variable to 32bits
+** \param[in] val - 16b value to convert
+** \return Converted value
+**/
+__INLINE DWORD INLINE__ conv16to32Bits(const WORD val) {
+	return (DWORD) (LSHIFT(val, 16) + val); }
+
+
+/*!\brief convert 64bits variable to 32bits
+** \param[in] val - 64b value to convert
+** \return Converted value
+**/
+__INLINE DWORD INLINE__ conv64to32Bits(const LWORD val) {
+	return (DWORD) RSHIFT64(val, 32); }
+
+/*!\brief convert 32bits variable to 64bits
+** \param[in] val - 32b value to convert
+** \return Converted value
+**/
+__INLINE LWORD INLINE__ conv32to64Bits(const DWORD val) {
+	return (LWORD) (LSHIFT64(val, 32) + val); }
+
+
+/*!\brief convert 8bits variable to 8+nb bits (16bits max)
 ** \warning conversion output shall not exceed 16bits (input shall strictly be unsigned 8bits)
 ** \warning nb shall be in range 0-8 (note that using 0 doesn't change val)
 ** \param[in] val - 8b value to convert
@@ -195,9 +225,9 @@ __INLINE WORD INLINE__ conv8to16Bits(const BYTE val) {
 ** \return Converted value
 **/
 __INLINE WORD conv8upto16Bits(const BYTE val, const BYTE nb) {
-	return ((WORD) ((WORD) (val << nb) + (WORD) (val & (0xFFU >> (8 - nb))))); }
+	return (WORD) ((val << nb) + (val >> (8 - nb))); }
 
-/*!\brief converts 16bits to 16+nb bits (32bits max)
+/*!\brief convert 16bits variable to 16+nb bits (32bits max)
 ** \warning conversion output shall not exceed 32bits (input shall strictly be unsigned 16bits)
 ** \warning nb shall be in range 0-16 (note that using 0 doesn't change val)
 ** \param[in] val - 16b value to convert
@@ -205,9 +235,9 @@ __INLINE WORD conv8upto16Bits(const BYTE val, const BYTE nb) {
 ** \return Converted value
 **/
 __INLINE DWORD conv16upto32Bits(const WORD val, const BYTE nb) {
-	return ((DWORD) ((DWORD) (val << nb) + (DWORD) (val & (0xFFFFU >> (16 - nb))))); }
+	return (DWORD) ((val << nb) + (val >> (16 - nb))); }
 
-/*!\brief converts 32bits to 32+nb bits (64bits max)
+/*!\brief convert 32bits variable to 32+nb bits (64bits max)
 ** \warning conversion output shall not exceed 64bits (input shall strictly be unsigned 32bits)
 ** \warning nb shall be in range 0-32 (note that using 0 doesn't change val)
 ** \param[in] val - 32b value to convert
@@ -215,7 +245,24 @@ __INLINE DWORD conv16upto32Bits(const WORD val, const BYTE nb) {
 ** \return Converted value
 **/
 __INLINE LWORD conv32upto64Bits(const DWORD val, const BYTE nb) {
-	return ((LWORD) ((LWORD) (val << nb) + (LWORD) (val & (0xFFFFFFFFUL >> (32 - nb))))); }
+	return (LWORD) (((LWORD) val << nb) + (val >> (32 - nb))); }
+
+
+/*!\brief convert x bits variable to y bits (32bits max)
+** \warning conversion output shall not exceed 32bits (input shall strictly be unsigned 32bits)
+** \warning nb shall be in range 0-32
+** \param[in] val - x bits value to convert
+** \param[in] from - number of bits from former variable
+** \param[in] to - number of bits of output variable
+** \return Converted value
+**/
+__INLINE DWORD convXtoYBits(const DWORD val, const BYTE from, const BYTE to)
+{
+	const int8_t diff_bits = min(32, to) - min(32, from);
+	if (diff_bits > 0)		{ return (DWORD) ((val << diff_bits) + (val >> (from - diff_bits))); }
+	else if (diff_bits < 0)	{ return (DWORD) RSHIFT(val, diff_bits); }
+	return val;
+}
 
 
 /************************/